day06_rag_pipeline/requirements.txt
openai
faiss-cpu
numpy

